Started adding support for NATS usage messages, updating to qms 0.0.4, and reorganizing code
