+++ internal/controllers/users.go
@@ -3,10 +3,8 @@ package controllers
 import (
 	"fmt"
 	"net/http"
-	"time"
 
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
 
 	"github.com/cyverse-de/go-mod/pbinit"
 	"github.com/cyverse-de/p/go/qms"
@@ -285,152 +283,4 @@ func (s Server) UpdateUserPlan(ctx echo.Context) error {
 
 		return model.Success(ctx, "Success", http.StatusOK)
 	})
-}
-
-type Usage struct {
-	Username     string  `json:"username"`
-	ResourceName string  `json:"resource_name"`
-	UsageValue   float64 `json:"usage_value"`
-	UpdateType   string  `json:"update_type"`
-}
-
-// AddUsages adds or updates the usage record for a user, plan, and resource type.
-func (s Server) AddUsages(ctx echo.Context) error {
-	var (
-		err   error
-		usage Usage
-	)
-
-	log := log.WithFields(logrus.Fields{"context": "adding usage information"})
-
-	context := ctx.Request().Context()
-
-	// Extract and validate the request body.
-	if err = ctx.Bind(&usage); err != nil {
-		return model.Error(ctx, "invalid request body", http.StatusBadRequest)
-	}
-	if usage.Username == "" {
-		return model.Error(ctx, "invalid username", http.StatusBadRequest)
-	}
-	if usage.ResourceName == "" {
-		return model.Error(ctx, "invalid resource name", http.StatusBadRequest)
-	}
-	if usage.UsageValue < 0 {
-		return model.Error(ctx, "invalid usage value", http.StatusBadRequest)
-	}
-	if usage.UpdateType == "" {
-		return model.Error(ctx, "missing usage update type value", http.StatusBadRequest)
-	}
-
-	log.Debug("validated usage information")
-
-	log = log.WithFields(logrus.Fields{
-		"user":       usage.Username,
-		"resource":   usage.ResourceName,
-		"updateType": usage.UpdateType,
-		"value":      usage.UsageValue,
-	})
-
-	err = s.GORMDB.Transaction(func(tx *gorm.DB) error {
-		// Look up the currently active user plan, adding a default plan if one doesn't exist already.
-		userPlan, err := db.GetActiveUserPlan(context, tx, usage.Username)
-		if err != nil {
-			return model.Error(ctx, err.Error(), http.StatusInternalServerError)
-		}
-
-		log.Debugf("active plan is %s", userPlan.Plan.Name)
-
-		// Look up the resource type.
-		resourceType, err := db.GetResourceTypeByName(context, tx, usage.ResourceName)
-		if err != nil {
-			return model.Error(ctx, err.Error(), http.StatusInternalServerError)
-		}
-		if resourceType == nil {
-			return model.Error(ctx, fmt.Sprintf("resource type '%s' does not exist", usage.ResourceName), http.StatusBadRequest)
-		}
-
-		log.Debug("found resource type in database")
-
-		// Initialize the new usage record.
-		var newUsage = model.Usage{
-			Usage:          usage.UsageValue,
-			UserPlanID:     userPlan.ID,
-			ResourceTypeID: resourceType.ID,
-		}
-
-		// Verify that the update operation for the given update type exists.
-		updateOperation := model.UpdateOperation{Name: usage.UpdateType}
-		err = tx.WithContext(context).Debug().First(&updateOperation).Error
-		if err == gorm.ErrRecordNotFound {
-			return model.Error(ctx, "invalid update type", http.StatusBadRequest)
-		}
-		if err != nil {
-			return model.Error(ctx, err.Error(), http.StatusInternalServerError)
-		}
-
-		log.Debug("verified update operation from database")
-
-		// Determine the current usage, which should be zero if the usage record doesn't exist.
-		currentUsage := model.Usage{
-			UserPlanID:     userPlan.ID,
-			ResourceTypeID: resourceType.ID,
-		}
-		err = tx.WithContext(context).Debug().First(&currentUsage).Error
-		if err != nil && err != gorm.ErrRecordNotFound {
-			return model.Error(ctx, err.Error(), http.StatusInternalServerError)
-		}
-
-		log.Debugf("got the current usage of %f", currentUsage.Usage)
-
-		// Update the new usage based on the values in the request body.
-		switch usage.UpdateType {
-		case UpdateTypeSet:
-			newUsage.Usage = usage.UsageValue
-		case UpdateTypeAdd:
-			newUsage.Usage = currentUsage.Usage + usage.UsageValue
-		default:
-			msg := fmt.Sprintf("invalid update type: %s", usage.UpdateType)
-			return model.Error(ctx, msg, http.StatusBadRequest)
-		}
-
-		log.Debugf("calculated the new usage to be %f", newUsage.Usage)
-
-		// Either add the new usage record or update the existing one.
-		err = tx.WithContext(context).Debug().Clauses(clause.OnConflict{
-			Columns: []clause.Column{
-				{
-					Name: "user_plan_id",
-				},
-				{
-					Name: "resource_type_id",
-				},
-			},
-			UpdateAll: true,
-		}).Create(&newUsage).Error
-		if err != nil {
-			return model.Error(ctx, err.Error(), http.StatusInternalServerError)
-		}
-
-		log.Debug("added/updated the usage record in the database")
-
-		// Record the update in the database.
-		update := model.Update{
-			Value:             newUsage.Usage,
-			ValueType:         ValueTypeUsages,
-			ResourceTypeID:    resourceType.ID,
-			EffectiveDate:     time.Now(),
-			UpdateOperationID: updateOperation.ID,
-		}
-		err = tx.WithContext(context).Debug().Create(&update).Error
-		if err != nil {
-			return model.Error(ctx, err.Error(), http.StatusInternalServerError)
-		}
-
-		log.Debug("recorded the update in the databse")
-
-		// Return a response to the caller.
-		msg := fmt.Sprintf("successfully updated the usage for: %s", usage.Username)
-		return model.SuccessMessage(ctx, msg, http.StatusOK)
-	})
-	return err
 }
